Initial Release
This repository includes the majority of the files I created for use
with my Web Programming course during the 2014-2015 school year. I
tried to minimize the repetition of publicly available files such as
Bootstrap, jQuery, etc, so you might need to add these back in for some
files to work correctly.